Move AES-CMAC implementation under fipsmodule boundary
