+++ crypto/CMakeLists.txt
@@ -290,7 +290,6 @@ add_library(
   cipher_extra/e_rc4.c
   cipher_extra/e_tls.c
   cipher_extra/tls_cbc.c
-  cmac/cmac.c
   conf/conf.c
   cpu-aarch64-fuchsia.c
   cpu-aarch64-linux.c
@@ -548,7 +547,6 @@ if(BUILD_TESTING)
     chacha/chacha_test.cc
     cipher_extra/aead_test.cc
     cipher_extra/cipher_test.cc
-    cmac/cmac_test.cc
     compiler_test.cc
     conf/conf_test.cc
     constant_time_test.cc
@@ -568,6 +566,7 @@ if(BUILD_TESTING)
     evp/scrypt_test.cc
     fipsmodule/aes/aes_test.cc
     fipsmodule/bn/bn_test.cc
+    fipsmodule/cmac/cmac_test.cc
     fipsmodule/ec/ec_test.cc
     fipsmodule/ec/p256-nistz_test.cc
     fipsmodule/ecdsa/ecdsa_test.cc
