+++ crypto/fipsmodule/bcm.c
@@ -60,6 +60,7 @@
 #include "cipher/e_aes.c"
 #include "cipher/e_des.c"
 #include "cipher/e_aesccm.c"
+#include "cmac/cmac.c"
 #include "des/des.c"
 #include "dh/check.c"
 #include "dh/dh.c"
